Fix warnings

+++ plugin/seq/libff-mmap-semaphore.c
@@ -305,7 +305,7 @@ void ffmmap_init(ff_Pmmap p, const char *nm, long len) {
 #else
   off_t size = lseek(p->fd, 0, SEEK_END);    // seek to end of file
   p->isnew = (size == 0);
-  printf(" len %ld size %ld \n", len, size);
+  printf(" len %lld size %lld \n", (long long)len, (long long)size);
   if (size < len) {
     if (ftruncate(p->fd, len) == -1) {
       perror("ftruncate");
